Normalize all time units to milliseconds Add a javascript formatter to render duration in HH:MM:SS.mmm Remove TestType.timeUnits Add CuantoNG sample project to InitializationService Consolidate SQL upgrade files

+++ api/src/main/groovy/cuanto/api/TestOutcome.groovy
@@ -26,7 +26,7 @@ class TestOutcome {
 	TestCase testCase
 	String testResult
     String testOutput
-    BigDecimal duration
+    Long duration
     String owner
     Bug bug
     String note
